v3.0
游戏功能已经完善，只差BGM了

+++ Assets/Scripts/EndGame.cs
@@ -8,7 +8,7 @@ public class EndGame : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        Record record = new Record(Global.name, Global.score, Global.dateTime);
+        Record record = new Record(Global.name, (int)Global.score, Global.dateTime);
         Record.saveRecord(record);
     }
 }
